refactor file names and directories
